When building on OS-X with CMake version 2.8.0 with an explicit CMAKE_INSTALL_PREFIX definition, the header files don't get installed, unless I make this change.
It appears to me that CMake won't traverse into the api/ directory
since that directory does not terminate in *.h.  If you try to change
the pattern to api/*.h, then the api directory still doesn't match.
However, since only the "exported" headers are in the api directory,
we might as well just tell cmake to install it all.

I can only guess that this must be a bug with CMake version 2.8.0.