Add interface to parse JSON into a tree like structure.
While reading back JSON data from files, I found myself fiddling again
and again with the scanner-like (think: lex) interface of
<yajl/yajl_parse.h>.

I think that the best approach to my problem is to write a *parser*,
i.e. a piece of code that gets tokens from the scanner and builds a
parse tree out of it. So this is what I did.

This doesn't play well with parsing streams, naturally, but that's not
my use-case and the stream parsing isn't affected by an additional
interface.

Please keep in mind that this is work in progress. It's currently not
even possible to free the returned structure. I just wanted to push it
out there in the hope of some early feedback.