src/yajl_tree.c: Parse JSON numbers to int and double and include that in the struct.
An additional "flags" field marks whether or not any problems occurred when
parsing the string. The check is very picky and even considers trailing
characters to be a problem. Especially when expecting an integer but the
string actually contains a double, the initial characters may form a valid
integer which differs greatly from the actual value of the double.