encode: don't check specific cases when the current char is large
Most characters don't get escaped.  According to the RFC, the only
character greater than '/' that gets escaped is '/'. Test for this
common case first, and only if the char is a candidate for escaping
check for the other escape cases.

This cuts an additional 29% off the runtime of yajl_string_encode