Fix paths in run_tests.sh helper scripts
When test/parsing/run_tests.sh was moved down a level from
its original location test/run_tests.sh, the paths it uses
were not updated. It is trying to access ../build/test
instead of ../../build/test.

The newly added test/api/run_tests.sh was not even looking
in the build directory, instead looking for executable files
in the current directory. This made it find itself, causing
it to exec itself resulting a fork bomb.

Signed-off-by: Daniel P. Berrange <berrange@redhat.com>